pyproject.toml
[tool.poetry]
name = "store_backend"
version = "0.1.0"
description = "A practice GraphQL backend."
authors = ["David Martinez <d@vidm.co>"]

[tool.poetry.dependencies]
python = "^3.7"
graphene = "^2.1"
sqlalchemy = "^1.3"
fastapi = "^0.46.0"
gunicorn = "^20.0"
uvicorn = "^0.11.1"
graphene_sqlalchemy = "^2.2"
python_dotenv = "^0.10.3"
mysqlclient = "^1.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
